Double Finger kind of working
If we move the Robot with Torques, we need to implement a torque -> pos
controller, which turns out to be hard to do for the panda.

If we control the joints directly by setting qpos or qvel, the contact
physics start to break down.

Solution: Have two fingers: One 2d finger floating through space pushing
the box, one finger attached to the robot that is controlled using IK
and qpos and following the first finger. Only the first one does
collisions.

This commit implements it, but has the roles of the fingers switched.

TODO:
- Make the magic finger do the collisions
- Stop the initial "leap" of the robot